model and optimization parameters

+++ includes/general_utils.cuh
@@ -49,10 +49,10 @@ inline torch::Tensor ImageToTorch(torch::Tensor image, std::vector<int64_t> reso
 struct Expon_lr_func {
     double lr_init;
     double lr_final;
-    int64_t lr_delay_steps;
+    double lr_delay_steps;
     double lr_delay_mult;
     int64_t max_steps;
-    Expon_lr_func(double lr_init, double lr_final, int64_t lr_delay_steps = 0, double lr_delay_mult = 1.0, int64_t max_steps = 1000000)
+    Expon_lr_func(double lr_init, double lr_final, double lr_delay_steps = 0, double lr_delay_mult = 1.0, int64_t max_steps = 1000000)
         : lr_init(lr_init),
           lr_final(lr_final),
           lr_delay_steps(lr_delay_steps),
@@ -64,12 +64,12 @@ struct Expon_lr_func {
             return 0.0;
         }
         double delay_rate;
-        if (lr_delay_steps > 0 && step != 0) {
-            delay_rate = lr_delay_mult + (1 - lr_delay_mult) * std::sin(0.5 * M_PI * std::min((double)step / (double)lr_delay_steps, 1.0));
+        if (lr_delay_steps > 0. && step != 0) {
+            delay_rate = lr_delay_mult + (1 - lr_delay_mult) * std::sin(0.5 * M_PI * std::min((double)step / lr_delay_steps, 1.0));
         } else {
             delay_rate = 1.0;
         }
-        double t = std::min(step / static_cast<double>(max_steps), 1.0);
+        double t = std::min(static_cast<double>(step) / static_cast<double>(max_steps), 1.0);
         double log_lerp = std::exp(std::log(lr_init) * (1 - t) + std::log(lr_final) * t);
         return delay_rate * log_lerp;
     }
